// src/components/marketplace/types.ts
export type ProgramCardModel = {
  category: string;
  coach: string;
  description: string;
  duration: string;
  id: string;
  image: string;
  level: string;
  price: string;
  title: string;
};

export type CoachCardModel = {
  discipline: string;
  id: string;
  image: string;
  name: string;
  signal: string;
};